requirements.txt
requests
beautifulsoup4
sentence-transformers
selenium
webdriver-manager
pywebview
cef-capi-py
